Move /cloud/user and /cloud/capabilities over to AppFramework
