+++ ocs/routes.php
@@ -81,21 +81,6 @@ API::register(
 	'core',
 	API::USER_AUTH
 	);
-// cloud
-API::register(
-	'get',
-	'/cloud/capabilities',
-	['OC_OCS_Cloud', 'getCapabilities'],
-	'core',
-	API::USER_AUTH
-	);
-API::register(
-	'get',
-	'/cloud/user',
-	['OC_OCS_Cloud', 'getCurrentUser'],
-	'core',
-	API::USER_AUTH
-);
 
 // Server-to-Server Sharing
 if (\OC::$server->getAppManager()->isEnabledForUser('files_sharing')) {
